feat(ui): split up StagingAreaToolbar
